be/services/gameServices/requirements.txt
flask
flask_sqlalchemy
flask_cors
psycopg2-binary
redis
python-dotenv
requests
pyjwt
